Proposed pay table edit
Proposed change to the salary table to make it really clear that the base compensation is not the whole story. I would like to ensure that nobody quickly glances at this table and thinks $133,444 is the top salary, when in fact $160,300 is the top salary (depending on locality). 

I would like to see user testing on this page eventually, of course.