Removed incorrect statement about Unpaid Leave
The Unpaid Leave section said that Parental Leave could only be taken if you've been employed by 18F for 12 months.  This is an incorrect statement.  You have to be a Federal employee for 12 months to qualify for _FMLA_ leave, which is not the same as parental leave.  FMLA leave is the preferred type of parental leave, but not the only leave option available to new parents.